fix: RenjuRule에서 List가 갱신되지 않던 문제 수정
