+++ src/main/kotlin/domain/BlackPlayer.kt
@@ -2,7 +2,7 @@ package domain
 
 class BlackPlayer(override val color: Color = Color.BLACK) : Player() {
     override fun validateRenju(stones: Stones, stone: Stone): Boolean {
-        return !stones.renjuRule.isThreeToThree(stone) && !stones.renjuRule.isFourToFour(stone) && stones.findScore(stone) < LARGE_PLACE
+        return !stones.renjuRule.isThreeToThree(stone) && !stones.renjuRule.isFourToFour(stone) && stones.renjuRule.findScore(stone) < LARGE_PLACE
     }
 
     companion object {
