+++ src/main/kotlin/domain/RenjuRule.kt
@@ -1,14 +1,15 @@
 package domain
 
-class RenjuRule(val stones: List<Stone>) {
+class RenjuRule(val stones: Stones) {
     fun isThreeToThree(stone: Stone): Boolean {
-        return Stones.directions.sumOf { direction ->
+        val v = directions.sumOf { direction ->
             checkOpenFourForLine((direction * -4) + stone.coordinate.point, direction)
-        } >= RENJU_LINE_CONDITION
+        }
+        return v >= RENJU_LINE_CONDITION
     }
 
     fun isFourToFour(stone: Stone): Boolean {
-        return Stones.directions.sumOf { direction ->
+        return directions.sumOf { direction ->
             checkFourForLine((direction * -4) + stone.coordinate.point, direction)
         } >= RENJU_LINE_CONDITION
     }
@@ -29,7 +30,7 @@ class RenjuRule(val stones: List<Stone>) {
         if (!validateCheckBlock(start, direction, OPEN_FOUR_BLOCK_SIZE)) return false
 
         val isOpened =
-            stones.none { it.coordinate.point == start } && stones.none { it.coordinate.point == (start + (direction * OPEN_FOUR_BLOCK_END)) }
+            stones.value.none { it.coordinate.point == start } && stones.value.none { it.coordinate.point == (start + (direction * OPEN_FOUR_BLOCK_END)) }
         val isFour = OPEN_FOUR_INNER_BLOCK_SHIFTED_RANGE.count {
             isBlackStoneThere(start + (direction * it))
         } == OPEN_FOUR_BLACK_STONE_COUNT
@@ -53,10 +54,29 @@ class RenjuRule(val stones: List<Stone>) {
     }
 
     private fun isBlackStoneThere(start: Point): Boolean {
-        val stone = stones.find { it.coordinate.point == start }
+        val stone = stones.value.find { it.coordinate.point == start }
         if (stone == null || stone.color == Color.WHITE) return false
         else if (stone.color == Color.BLACK) return true
         return false
+    }
+
+    fun findScore(stone: Stone): Int {
+        return directions.maxOf { item ->
+            val nextCoordinate = (stone.coordinate + item) ?: return@maxOf 0
+            val score = startSearch(nextCoordinate, item, stone.color, INITIAL_SCORE)
+            val invertedDirections = item * -1
+            val invertCoordinate = (stone.coordinate + invertedDirections) ?: return@maxOf 0
+            val invertedScore = startSearch(invertCoordinate, invertedDirections, stone.color, INITIAL_SCORE)
+            score + invertedScore
+        }
+    }
+
+    private fun startSearch(coordinate: Coordinate, direction: Point, color: Color, count: Int): Int {
+        if (stones.value.any { it.coordinate == coordinate && it.color == color }) {
+            val nextCoordinate = (coordinate + direction) ?: return count + SEARCH_INTERVAL
+            return startSearch(nextCoordinate, direction, color, count + SEARCH_INTERVAL)
+        }
+        return count
     }
 
     companion object {
@@ -69,5 +89,12 @@ class RenjuRule(val stones: List<Stone>) {
         private const val OPEN_FOUR_BLOCK_END = 5
         private const val OPEN_FOUR_BLACK_STONE_COUNT = 2
         private const val FOUR_BLACK_STONE_COUNT = 3
+
+        private const val INITIAL_SCORE = 0
+        private const val SEARCH_INTERVAL = 1
+
+        val directions = listOf(
+            Point(-1, 1), Point(0, 1), Point(1, 1), Point(1, 0)
+        )
     }
 }
